the_bank_app/requirements.txt
# Synthetic Retail Bank - Python Dependencies
# Updated for Python 3.12 compatibility

# Core framework
streamlit>=1.40.0

# Snowflake connectivity
snowflake-snowpark-python>=1.40.0
snowflake-connector-python>=3.12.0

# Data manipulation
pandas>=2.2.0
numpy>=1.26.0

# Visualization
plotly>=5.24.0

# Utilities
python-dateutil>=2.9.0
requests>=2.31.0
urllib3>=2.2.0
